+++ R/utils.R
@@ -662,3 +662,14 @@ vcov_gee <- function(obj) {
     out <- obj$geese$vbeta
     return(out)
 }
+
+#' Find vcov for quantile regression models
+#'
+#' @param obj a \code{rq} class object.
+
+vcov_rq <- function(obj) {
+    if (!("rq" %in% class(obj)))
+        stop('Not an rq class object', call. = FALSE)
+    out <- summary(obj, cov = TRUE)$cov
+    return(out)
+}