+++ R/wrappers.R
@@ -19,38 +19,61 @@
 #'   effects in the form \code{x1*x2} without computing them in prior
 #'   steps; \code{I(x1*x2)} to include only the interaction term and
 #'   exclude the main effects; and quadratic terms in the form
-#'   \code{I(x1^2)}
-#' @param model the name of a statistical model.
-#'   For a list of supported models visit:
+#'   \code{I(x1^2)}.
+#' @param model the name of a statistical model to estimate.
+#'   For a list of supported models and their documentation see:
 #'   \url{http://docs.zeligproject.org/articles/}.
 #' @param data the name of a data frame containing the variables
-#'   referenced in the formula, or a list of multiply imputed data frames
+#'   referenced in the formula or a list of multiply imputed data frames
 #'   each having the same variable names and row numbers (created by
-#'   \code{mi})
+#'   \code{Amelia} or \code{\link{to_zelig_mi}}).
 #' @param ... additional arguments passed to \code{zelig},
-#'   depending on the model to be estimated
-#' @param by a factor variable contained in \code{data}. Zelig will subset
+#'   relevant for the model to be estimated.
+#' @param by a factor variable contained in \code{data}. If supplied,
+#'   \code{zelig} will subset
 #'   the data frame based on the levels in the \code{by} variable, and
-#'   estimate a model for each subset. This a particularly powerful option
-#'   which will allow you to save a considerable amount of effort. For
-#'   example, to run the same model on all fifty states, you could type:
-#'   \code{z.out <- zelig(y ~ x1 + x2, data = mydata, model = 'ls', by = 'state')}
-#'   You may also use \code{by} to run models using MatchIt subclass
-#' @param cite If is set to 'TRUE' (default), the model citation will be
+#'   estimate a model for each subset. This can save a considerable amount of
+#'   effort. For example, to run the same model on all fifty states, you could
+#'   use: \code{z.out <- zelig(y ~ x1 + x2, data = mydata, model = 'ls',
+#'   by = 'state')} You may also use \code{by} to run models using MatchIt
+#'   subclasses.
+#' @param cite If is set to 'TRUE' (default), the model citation will be printed
+#'   to the console.
+#'
+#' @details
+#' Additional parameters avaialable to many models include:
+#' \itemize{
+#'   \item weights: vector of weight values or a name of a variable in the dataset
+#'   by which to weight the model. For more information see:
+#'   \url{http://docs.zeligproject.org/articles/weights.html}.
+#'   \item bootstrap: logical or numeric. If \code{FALSE} don't use bootstraps to
+#'   robustly estimate uncertainty around model parameters due to sampling error.
+#'   If an integer is supplied, the number of boostraps to run.
+#'   For more information see:
+#'   \url{http://docs.zeligproject.org/articles/bootstraps.html}.
+#' }
+
+#'
 #' @return Depending on the class of model selected, \code{zelig} will return
 #'   an object with elements including \code{coefficients}, \code{residuals},
 #'   and \code{formula} which may be summarized using
 #'   \code{summary(z.out)} or individually extracted using, for example,
-#'   \code{z.out\$coefficients}. See the specific models listed above
-#'   for additional output values, or simply type \code{names(z.out)}.
+#'   \code{coef(z.out)}. See
+#'   \url{http://docs.zeligproject.org/articles/getters.html} for a list of
+#'   functions to extract model components. You can also extract whole fitted
+#'   model objects using \code{\link{from_zelig_model}}.
 #'
+#' @seealso \url{http://docs.zeligproject.org/articles/}
 #' @name zelig
 #' @author Matt Owen, Kosuke Imai, Olivia Lau, and Gary King
-#' Maintainer: Christopher Gandrud
-#' @keywords package
 #' @export
 
-zelig <- function(formula, model, data, ..., by = NULL, cite = TRUE) {
+zelig <- function(formula,
+                  model,
+                  data,
+                  ...,
+                  by = NULL,
+                  cite = TRUE) {
     # .Deprecated('\nz$new() \nz$zelig(...)') Check if required model argument is
     # specified
     if (missing(model))
@@ -158,7 +181,9 @@ zelig <- function(formula, model, data, ..., by = NULL, cite = TRUE) {
 
 setx <- function(obj, fn = NULL, data = NULL, cond = FALSE, ...) {
     # .Deprecated('\nz$new() \nz$zelig(...) \nz$setx() or z$setx1 or z$setrange')
-    is_zelig(obj)
+
+    if(!is_zelig(obj, fail = FALSE))
+        obj <- to_zelig(obj)
 
     x5 <- obj$copy()
     # This is the length of each argument in '...'s