+++ tests/testthat/test-logit.R
@@ -1,5 +1,4 @@
 # REQUIRE TEST Monte Carlo test logit ------------------------------------------
-
 test_that('REQUIRE TEST logit Monte Carlo', {
     z <- zlogit$new()
     test <- z$mcunit(minx = -2, maxx = 2, plot = FALSE)
@@ -17,4 +16,14 @@ test_that('REQUIRE TEST logit example and show odds_ratios', {
     ors <- ors$summ[[1]]$coefficients[1:3]
 
     expect_equal(exp(betas)[[1]], ors[1])
+})
+
+# REQUIRE TEST logit example and show odds_ratios ------------------------------
+test_that('REQUIRE TEST logit to_zelig', {
+    data(turnout)
+    m1 <- glm(vote ~ age + race, family = binomial(link="logit"),
+              data = turnout)
+    m1_sims <- sim(setx(m1))
+    expect_equal(sort(unique(zelig_qi_to_df(m1_sims)$predicted_value)), c(0, 1))
+
 })